Crate debruijn

debruijn-rs: a De Bruijn graph for DNA seqeunces in Rust.

This library provides tools for efficiently construction de bruijn graphs from DNA sequences, tracking arbitrary metadata associated with kmers in the graph, and performing path-compression of unbranched graph paths to improve speed and reduce memory consumption. All the data structures in debruijn-rs are specialized to the alphabet {'A', 'C', 'G', 'T'}, and use 2-bit packed encoding of base-pairs into integer types, and efficient methods for reverse complement, enumerating kmers from longer sequences, and transfering data between sequences.

Modules

dna_string
A 2-bit encoding of arbitrary length DNA sequences.

filter
Methods for filtering observed kmers before De Bruijn graph construction, and summarizing 'color' annotations.

kmer
Represent kmers with statically know length in compact integer types

msp
Methods for minimum substring partitioning of a DNA string

paths
Compute path-compressed De Bruijn graphs from kmers or intermediate sized De Bruijn graph fragments

vmer
Variable-length DNA strings packed into fixed-size structs.

Structs

Exts
Store single-base extensions for a DNA Debruijn graph.

KmerExtsIter
Iterate over the (Kmer, Exts) tuples of a sequence and it's extensions efficiently

KmerIter
Iterate over the Kmers of a DNA sequence efficiently

Enums

Dir
Direction of motion in a DeBruijn graph

Traits

Kmer
Encapsulates a Kmer sequence with statically known K.

Mer
Generic trait for interacting with DNA sequences

MerImmut
An immutable interface to a Mer sequence.

Vmer
A DNA sequence with run-time variable length, up to a statically known maximum length

Functions

base_to_bits
Convert an ASCII-encoded DNA base to a 2-bit representation

bits_to_ascii
Convert a 2-bit representation of a base to a char

bits_to_base
Convert a 2-bit representation of a base to a char

complement
The complement of a 2-bit encoded base