Trait debruijn::compression::CompressionSpec

pub trait CompressionSpec<D> {
    fn reduce(&self, path_object: D, kmer_object: &D) -> D;

    fn join_test(&self, d1: &D, d2: &D) -> bool;
}

Customize the path-compression process. Implementing this trait lets the user control how the per-kmer data (of type D) is summarized into a per-path summary data (of type DS). It also let's the user introduce new breaks into paths by inspecting in the per-kmer data of a proposed with join_test_kmer function.

Required methods

fn reduce(&self, path_object: D, kmer_object: &D) -> D

fn join_test(&self, d1: &D, d2: &D) -> bool

Loading content...