Struct debruijn::dna_string::DnaString

pub struct DnaString { /* fields omitted */ }

A container for sequence of DNA bases.

Methods

impl DnaString

pub fn new() -> DnaString

Create an empty DNA string

pub fn len(&self) -> usize

Length of the sequence

pub fn with_capacity(n: usize) -> Self

Create a new instance with a given capacity.

pub fn blank(n: usize) -> Self

Create a DnaString of length n initialized to all A's

pub fn from_dna_string(dna: &str) -> DnaString

Create a DnaString corresponding to an ACGT-encoded str.

pub fn from_dna_only_string(dna: &str) -> Vec<DnaString>

Create a DnaString corresponding to an ACGT-encoded str.

pub fn from_acgt_bytes(bytes: &[u8]) -> DnaString

Create a DnaString from an ASCII ACGT-encoded byte slice. Non ACGT positions will be converted to 'A'

pub fn from_acgt_bytes_hashn(bytes: &[u8], read_name: &[u8]) -> DnaString

Create a DnaString from an ACGT-encoded byte slice, Non ACGT positions will be converted to repeatable random base determined by a hash of the read name and the position within the string.

pub fn from_bytes(bytes: &[u8]) -> DnaString

Create a DnaString from a 0-4 encoded byte slice

pub fn to_string(&self) -> String

Convert sequence to a String

pub fn to_bytes(&self) -> Vec<u8>

Convert sequence to a Vector of 0-4 encoded bytes

pub fn to_ascii_vec(&self) -> Vec<u8>

Convert sequence to a Vector of ascii-encoded bytes

pub fn push(&mut self, value: u8)

Append a 0-4 encoded base.

pub fn push_bytes(&mut self, bytes: &Vec<u8>, seq_length: usize)

Push 0-4 encoded bases from a byte array.

Arguments

bytes: byte array to read values from seq_length: how many values to read from the byte array. Note that this is number of values not number of elements of the byte array.

Important traits for DnaStringIter<'a>
impl<'a> Iterator for DnaStringIter<'a> type Item = u8;

pub fn iter(&self) -> DnaStringIter

Iterate over stored values (values will be unpacked into bytes).

pub fn clear(&mut self)

Clear the sequence.

pub fn is_empty(&self) -> bool

pub fn prefix(&self, k: usize) -> DnaStringSlice

Get the length k prefix of the DnaString

pub fn suffix(&self, k: usize) -> DnaStringSlice

Get the length k suffix of the DnaString

pub fn slice(&self, start: usize, end: usize) -> DnaStringSlice

Get slice containing the interval [start, end) of self

pub fn reverse(&self) -> DnaString

Create a fresh DnaString containing the reverse of self