Struct debruijn::dna_string::PackedDnaStringSet

pub struct PackedDnaStringSet {
    pub sequence: DnaString,
    pub start: Vec<usize>,
    pub length: Vec<u32>,
}

Container for many distinct sequences, concatenated into a single DnaString. Each sequence is accessible by index as a DnaStringSlice.

Fields
sequence: DnaString start: Vec<usize> length: Vec<u32>
Methods

impl<'a> PackedDnaStringSet

pub fn new() -> Self

Create an empty PackedDnaStringSet

pub fn get(&'a self, i: usize) -> DnaStringSlice<'a>

Get a DnaStringSlice containing ith sequence in the set

pub fn slice(&'a self, i: usize, start: usize, end: usize) -> DnaStringSlice<'a>

Get a DnaStringSlice containing ith sequence in the set

pub fn len(&self) -> usize

Number of sequences in the set

pub fn add<'b, R: Borrow<u8>, S: IntoIterator<Item = R>>(&mut self, sequence: S)