Module debruijn::filter

Methods for filtering observed kmers before De Bruijn graph construction, and summarizing 'color' annotations.

Structs

CountFilterSet

Traits

KmerSummarizer

Functions

filter_kmers
Read a shard and determine the valid kmers Low memory implementation that should consume < 4G of temporary memory To reduce memory consumption, set track_bcs to false to forget about BC lists.