Struct debruijn::graph::BaseGraph

pub struct BaseGraph<K, D> {
    pub sequences: PackedDnaStringSet,
    pub exts: Vec<Exts>,
    pub data: Vec<D>,
    pub stranded: bool,
    // some fields omitted
}

A compressed DeBruijn graph carrying auxiliary data on each node of type D. This type does not carry the sorted index arrays the allow the graph to be walked efficiently. The DeBruijnGraph type wraps this type and add those vectors.

Fields
sequences: PackedDnaStringSet exts: Vec<Exts> data: Vec<D> stranded: bool
Methods

impl<K, D> BaseGraph<K, D>

pub fn new(stranded: bool) -> Self

pub fn len(&self) -> usize

pub fn combine<I: Iterator<Item = BaseGraph<K, D>>>(graphs: I) -> Self

impl<K: Kmer, D> BaseGraph<K, D>

pub fn add<'b, R: Borrow<u8>, S: IntoIterator<Item = R>>(
&mut self,
sequence: S,
exts: Exts,
data: D
)

pub fn finish(self) -> DebruijnGraph<K, D>