Struct debruijn::graph::Node

pub struct Node<'a, K: Kmer + 'a, D: 'a> {
    pub node_id: usize,
    pub graph: &'a DebruijnGraph<K, D>,
}

Unbranched sequence in the DeBruijn graph

Fields
node_id: usize graph: &'a DebruijnGraph<K, D>
Methods

impl<'a, K: Kmer, D: Debug> Node<'a, K, D>

pub fn len(&self) -> usize

Length of the sequence of this node

pub fn sequence(&self) -> DnaStringSlice<'a>

Sequence of the node

pub fn data(&self) -> &'a D

Reference to auxiliarly data associated with the node

pub fn exts(&self) -> Exts

Extension bases from this node

pub fn l_edges(&self) -> SmallVec<[(usize, Dir, bool); 4]>

Edges leaving the left side of the node in the format

pub fn r_edges(&self) -> SmallVec<[(usize, Dir, bool); 4]>

Edges leaving the right side of the node in the format

pub fn edges(&self, dir: Dir) -> SmallVec<[(usize, Dir, bool); 4]>

Edges leaving the 'dir' side of the node in the format