Trait debruijn::Kmer

pub trait Kmer: Mer + Sized + Copy + PartialEq + PartialOrd + Eq + Ord + Hash {
    fn empty() -> Self;
    fn k() -> usize;

    fn min_rc_flip(&self) -> (Self, bool) { ... }
    fn min_rc(&self) -> Self { ... }
    fn is_palindrome(&self) -> bool { ... }
    fn from_bytes(bytes: &[u8]) -> Self { ... }
    fn from_ascii(bytes: &[u8]) -> Self { ... }
    fn to_string(&self) -> String { ... }
    fn kmers_from_string(str: &[u8]) -> Vec<Self> { ... }
}

Encapsulates a Kmer sequence with statically known K.

Required Methods

fn empty() -> Self

Create a Kmer initialized to all A's

fn k() -> usize

K value for this concrete type.

Provided Methods

fn min_rc_flip(&self) -> (Self, bool)

Return the minimum of the kmer and it's reverse complement, and a flag indicating if sequence was flipped

fn min_rc(&self) -> Self

fn is_palindrome(&self) -> bool

Test if this Kmer and it's reverse complement are the same

fn from_bytes(bytes: &[u8]) -> Self

Create a Kmer from the first K bytes of bytes

fn from_ascii(bytes: &[u8]) -> Self

fn to_string(&self) -> String

fn kmers_from_string(str: &[u8]) -> Vec<Self>

Generate all kmers from string