Struct debruijn::dna_string::DnaString

pub struct DnaString { /* fields omitted */ }

A sequence of DnaStringoded values. This implementation is specialized to 2-bit alphabets

Methods

impl DnaString

fn new() -> DnaString

Create a new instance with a given encoding width (e.g. width=2 for using two bits per value).

fn empty(n: usize) -> Self

Create a new instance with a given capacity.

fn from_dna_string(dna: &str) -> DnaString

fn from_bytes(bytes: &Vec<u8>) -> DnaString

fn to_dna_string(&self) -> String

fn to_ascii_vec(&self) -> Vec<u8>

fn push(&mut self, value: u8)

Append a value.

fn push_bytes(&mut self, bytes: &Vec<u8>, seq_length: usize)

Push values read from a byte array.

Arguments

bytes: byte array to read values from seq_length: how many values to read from the byte array. Note that this is number of values not number of elements of the byte array.

fn push_values(&mut self, n: usize, value: u8)

Append n times the given value.

fn iter(&self) -> DnaStringIter

Iterate over stored values (values will be unpacked into bytes).

fn clear(&mut self)

Clear the sequence.

fn width(&self) -> usize

fn is_empty(&self) -> bool

fn pref(&self, k: usize) -> DnaString

Get the prefix of a given size.

fn suf(&self, k: usize) -> DnaString

fn reverse(&self) -> DnaString