Trait debruijn::Mer

pub trait Mer: Sized + Debug {
    fn len(&self) -> usize;

    fn get(&self, pos: usize) -> u8;

    fn set_mut(&mut self, pos: usize, val: u8);

    fn set_slice_mut(&mut self, pos: usize, nbases: usize, value: u64);

    fn rc(&self) -> Self;

    fn iter<'a>(&'a self) -> MerIter<'a, Self> { ... }
}

Trait for interacting with DNA sequences

Required methods

fn len(&self) -> usize

Length of DNA sequence

fn get(&self, pos: usize) -> u8

Get 2-bit encoded base at position pos

fn set_mut(&mut self, pos: usize, val: u8)

Set base at pos to 2-bit encoded base val

fn set_slice_mut(&mut self, pos: usize, nbases: usize, value: u64)

Set nbases positions in the sequence, starting at pos. Values must be packed into the upper-most bits of value.

fn rc(&self) -> Self

Return a new object containing the reverse complement of the sequence

Loading content...
Provided methods

Important traits for MerIter<'a, M>
impl<'a, M: 'a + Mer> Iterator for MerIter<'a, M> type Item = u8;

fn iter<'a>(&'a self) -> MerIter<'a, Self>

Iterate over the bases in the sequence

Loading content...