Trait debruijn::Vmer

pub trait Vmer: Mer + PartialEq + Eq {
    fn new(len: usize) -> Self;

    fn max_len() -> usize;

    fn get_kmer<K: Kmer>(&self, pos: usize) -> K;

    fn from_slice(seq: &[u8]) -> Self { ... }

    fn first_kmer<K: Kmer>(&self) -> K { ... }

    fn last_kmer<K: Kmer>(&self) -> K { ... }

    fn both_term_kmer<K: Kmer>(&self) -> (K, K) { ... }

    fn term_kmer<K: Kmer>(&self, dir: Dir) -> K { ... }

    fn iter_kmers<K: Kmer>(&self) -> KmerIter<K, Self> { ... }

    fn iter_kmer_exts<K: Kmer>(&self, seq_exts: Exts) -> KmerExtsIter<K, Self> { ... }
}

A DNA sequence with run-time variable length, up to a statically known maximum length

Required Methods

fn new(len: usize) -> Self

Create a new sequence with length len, initialized to all A's

fn max_len() -> usize

Maximum sequence length that can be stored in this type

fn get_kmer<K: Kmer>(&self, pos: usize) -> K

Efficiently extract a Kmer from the sequence

Provided Methods

fn from_slice(seq: &[u8]) -> Self

Create a Vmer from a sequence of bytes

fn first_kmer<K: Kmer>(&self) -> K

Get the first Kmer from the sequence

fn last_kmer<K: Kmer>(&self) -> K

Get the last kmer in the sequence

fn both_term_kmer<K: Kmer>(&self) -> (K, K)

Get the terminal kmer of the sequence, on the both side of the sequence

fn term_kmer<K: Kmer>(&self, dir: Dir) -> K

Get the terminal kmer of the sequence, on the side of the sequence given by dir

Important traits for KmerIter<'a, K, D>
impl<'a, K: Kmer, D: Mer> Iterator for KmerIter<'a, K, D> type Item = K;

fn iter_kmers<K: Kmer>(&self) -> KmerIter<K, Self>

Iterate over the kmers in the sequence

Important traits for KmerExtsIter<'a, K, D>
impl<'a, K: Kmer, D: Mer> Iterator for KmerExtsIter<'a, K, D> type Item = (K, Exts);

fn iter_kmer_exts<K: Kmer>(&self, seq_exts: Exts) -> KmerExtsIter<K, Self>

Iterate over the kmers and their extensions, given the extensions of the whole sequence